move plot tests to test.rs, keeping guards
