+++ rust/src/traits/samplers/discretize/mod.rs
@@ -151,66 +151,3 @@ impl_cast_internal_rational_int!(u8 u16 u32 u64 u128 usize i8 i16 i32 i64 i128 i
 
 #[cfg(test)]
 mod test;
-
-#[cfg(all(test, feature = "test-plot"))]
-mod test_plotting {
-    use super::*;
-    use crate::error::ExplainUnwrap;
-    use crate::traits::samplers::Fallible;
-
-    fn plot_continuous(title: String, data: Vec<f64>) -> Fallible<()> {
-        use vega_lite_4::*;
-
-        VegaliteBuilder::default()
-            .title(title)
-            .data(&data)
-            .mark(Mark::Area)
-            .transform(vec![TransformBuilder::default().density("data").build()?])
-            .encoding(
-                EdEncodingBuilder::default()
-                    .x(XClassBuilder::default()
-                        .field("value")
-                        .position_def_type(Type::Quantitative)
-                        .build()?)
-                    .y(YClassBuilder::default()
-                        .field("density")
-                        .position_def_type(Type::Quantitative)
-                        .build()?)
-                    .build()?,
-            )
-            .build()?
-            .show()
-            .unwrap_test();
-        Ok(())
-    }
-
-    #[test]
-    #[ignore] // Don't want to produce graphics in CI
-    fn plot_laplace() -> Fallible<()> {
-        let shift = 0.;
-        let scale = 5.;
-
-        let title = format!("Laplace(shift={}, scale={}) distribution", shift, scale);
-        let data = (0..10_000)
-            .map(|_| sample_discrete_laplace_Z2k(shift, scale, -1074))
-            .collect::<Fallible<Vec<f64>>>()?;
-
-        plot_continuous(title, data).unwrap_test();
-        Ok(())
-    }
-
-    #[test]
-    #[ignore] // Don't want to produce graphics in CI
-    fn plot_gaussian() -> Fallible<()> {
-        let shift = 0.;
-        let scale = 5.;
-
-        let title = format!("Gaussian(shift={}, scale={}) distribution", shift, scale);
-        let data = (0..10_000)
-            .map(|_| sample_discrete_gaussian_Z2k(shift, scale, -1074))
-            .collect::<Fallible<Vec<f64>>>()?;
-
-        plot_continuous(title, data).unwrap_test();
-        Ok(())
-    }
-}
