+++ rust/src/traits/samplers/discretize/test.rs
@@ -95,3 +95,67 @@ fn test_find_nearest_multiple_of_2k() -> Fallible<()> {
     );
     Ok(())
 }
+
+
+#[cfg(feature = "test-plot")]
+mod test_plotting {
+    use super::*;
+    use crate::error::ExplainUnwrap;
+    use crate::traits::samplers::Fallible;
+
+    fn plot_continuous(title: String, data: Vec<f64>) -> Fallible<()> {
+        use vega_lite_4::*;
+
+        VegaliteBuilder::default()
+            .title(title)
+            .data(&data)
+            .mark(Mark::Area)
+            .transform(vec![TransformBuilder::default().density("data").build()?])
+            .encoding(
+                EdEncodingBuilder::default()
+                    .x(XClassBuilder::default()
+                        .field("value")
+                        .position_def_type(Type::Quantitative)
+                        .build()?)
+                    .y(YClassBuilder::default()
+                        .field("density")
+                        .position_def_type(Type::Quantitative)
+                        .build()?)
+                    .build()?,
+            )
+            .build()?
+            .show()
+            .unwrap_test();
+        Ok(())
+    }
+
+    #[test]
+    #[ignore] // Don't want to produce graphics in CI
+    fn plot_laplace() -> Fallible<()> {
+        let shift = 0.;
+        let scale = 5.;
+
+        let title = format!("Laplace(shift={}, scale={}) distribution", shift, scale);
+        let data = (0..10_000)
+            .map(|_| sample_discrete_laplace_Z2k(shift, scale, -1074))
+            .collect::<Fallible<Vec<f64>>>()?;
+
+        plot_continuous(title, data).unwrap_test();
+        Ok(())
+    }
+
+    #[test]
+    #[ignore] // Don't want to produce graphics in CI
+    fn plot_gaussian() -> Fallible<()> {
+        let shift = 0.;
+        let scale = 5.;
+
+        let title = format!("Gaussian(shift={}, scale={}) distribution", shift, scale);
+        let data = (0..10_000)
+            .map(|_| sample_discrete_gaussian_Z2k(shift, scale, -1074))
+            .collect::<Fallible<Vec<f64>>>()?;
+
+        plot_continuous(title, data).unwrap_test();
+        Ok(())
+    }
+}
