+++ rust/src/traits/samplers/geometric/mod.rs
@@ -231,43 +231,5 @@ pub(super) fn sample_geometric_buffer(
     })
 }
 
-#[cfg(all(test, feature = "test-plot"))]
-mod test_plotting {
-    use super::*;
-    use crate::error::ExplainUnwrap;
-    use crate::traits::samplers::Fallible;
-    #[test]
-    #[ignore] // Don't want to produce graphics in CI
-    fn plot_geometric() -> Fallible<()> {
-        let shift = 0;
-        let scale = 5.;
-
-        let title = format!("Geometric(shift={}, scale={}) distribution", shift, scale);
-        let data = (0..10_000)
-            .map(|_| sample_discrete_laplace_linear(0, 1., None))
-            .collect::<Fallible<Vec<i8>>>()?;
-
-        use vega_lite_4::*;
-        VegaliteBuilder::default()
-            .title(title)
-            .data(&data)
-            .mark(Mark::Bar)
-            .encoding(
-                EdEncodingBuilder::default()
-                    .x(XClassBuilder::default()
-                        .field("data")
-                        .position_def_type(Type::Nominal)
-                        .build()?)
-                    .y(YClassBuilder::default()
-                        .field("data")
-                        .position_def_type(Type::Quantitative)
-                        .aggregate(NonArgAggregateOp::Count)
-                        .build()?)
-                    .build()?,
-            )
-            .build()?
-            .show()
-            .unwrap_test();
-        Ok(())
-    }
-}
+#[cfg(test)]
+mod test;
